proper realm quries
