+++ EZWorkout/ExcerciseTableViewCell.swift
@@ -1,10 +1,4 @@
-//
-//  ExcerciseTableViewCell.swift
-//  EZWorkout
-//
-//  Created by Yilei Huang on 2019-03-01.
-//  Copyright © 2019 Spencer Symington. All rights reserved.
-//
+
 
 import UIKit
